Issue #2333 Refactored satellite to store table values in CRS space instead of lat/lon for GOES-R readiness
Amend: Accepted ticket.


Former-commit-id: fe9573d1040f054763f8680426106898351806ae [formerly fe9573d1040f054763f8680426106898351806ae [formerly 780ac0129fe27b5ffb08a320f19820065f1cf415]] [formerly e3f6b1ea21e4f224bdde1fed2f5ff4bdcb86de9c] [formerly fe9573d1040f054763f8680426106898351806ae [formerly fe9573d1040f054763f8680426106898351806ae [formerly 780ac0129fe27b5ffb08a320f19820065f1cf415]] [formerly e3f6b1ea21e4f224bdde1fed2f5ff4bdcb86de9c] [formerly 2c9a7ea31990550418585c3d34a353e6ad63f245 [formerly 2c9a7ea31990550418585c3d34a353e6ad63f245 [formerly 866c851b47eec6e4d924c26c4f23ed8a150dde90]] [formerly e3f6b1ea21e4f224bdde1fed2f5ff4bdcb86de9c [formerly e3f6b1ea21e4f224bdde1fed2f5ff4bdcb86de9c [formerly 384b8e6a02d4fffa0688993b2a2b53e974ecf493]]]]]
Former-commit-id: 2c9a7ea31990550418585c3d34a353e6ad63f245 [formerly 2c9a7ea31990550418585c3d34a353e6ad63f245 [formerly 866c851b47eec6e4d924c26c4f23ed8a150dde90]]
Former-commit-id: 8a353aa6598436120e94efdd933cc62b7209f82a [formerly e30ca57b9e4e6ebbd8fe7977bd17abe28ac58249] [formerly bdbea7bbeb5e8e9880b6ba967269d07aeb5b328a [formerly 43b7cb3107322b5753008fac386fc28db01f9af0]] [formerly fef34a74307e432168ea955f5a7500aa20c64423 [formerly 01af177e48af9aa69310c25fc487a980deb46dc8] [formerly b03081056b8f3e9edb627191e45f11b1c45479ad [formerly ba25f56dbe9b401ac3bc3d6a3cda203bf366effd]]]
Former-commit-id: ef0e4291679d29bc105b4a1621140664b578edf4 [formerly a436dc7fa725ae9dd6a3a51636cd1495b804f8c4] [formerly 18f78df64c58d707d2a78835a4354c11e64383b2 [formerly 80efdbb2b4ca51399d3c591fb8793b52da2886cb]]
Former-commit-id: ecd8c1f9987e55eeef39302d4d8dd0d6425b53df [formerly 6e5873733e3a421458696db15fc4a2dc7e54e76f]
Former-commit-id: 683972ba9b5cc8a04804c0453d64fd488ef3c832

+++ edexOsgi/com.raytheon.uf.edex.plugin.fog/src/com/raytheon/uf/edex/plugin/fog/common/FogConfig.java
@@ -292,8 +292,7 @@ public class FogConfig {
      * @return
      */
     private void setDx(SatelliteRecord rec) {
-
-        dx = rec.getCoverage().getDx();
+        dx = (float) rec.getCoverage().getDx();
     }
 
     /**
@@ -312,8 +311,7 @@ public class FogConfig {
      * @return
      */
     private void setDy(SatelliteRecord rec) {
-
-        dy = rec.getCoverage().getDy();
+        dy = (float) rec.getCoverage().getDy();
     }
 
     /**
